+++ src/main/java/com/shapecraft/block/BlockPoolManager.java
@@ -78,5 +78,9 @@ public class BlockPoolManager {
         public boolean isDoor() {
             return "door".equals(blockType);
         }
+
+        public boolean isTrapdoor() {
+            return "trapdoor".equals(blockType);
+        }
     }
 }